net: move BanMan to its own files
